Move normalize to util
